Factor out CircuitStateController
Refactor to pull common circuit state functionality in to
CircuitStateController.  More clearly exposes the function of individual
variants, Frequency/ Proportion/ ConsecutiveCount/ Timeslice/
controllers.
